Test Program one
I know there may be better ways of doing this code but this is using most of the stuff i've learnt from my A.S classes, Im posting this here for a member of my class to look at. There at a few errors at the minute but im working on them, and adding functionality step by step, As this is just a proof of concept not anything like I will be putting into a final program for use in the game i am wanting to create. Thank you in advance for if there are any glaringly obvious mistakes ive made which could be contributing to my errors when trying to read from the file and only being able to populate positions 0 through 3 (4 accounts)in the arrays instead of the 6 account files I have saved I would appreciate the feed back but please try not going anything too advanced as we are limited with the knowledge we need to know to pass the first year exams, if you can tell me both ways; the advanced and in simpler form that would be great