Create TestProgram(additional features and new bug for MrlukeKay to look at)
i think there might be a bug in the WriteExistingUserData()sub routine. its not updating the file even though the routine is being called also only 2 positons are being populated by the arrays even though ive added multiple it saves the files out with all the correct data? also luke ive added the formatting for date of birth and a error capture to make them enter again if they got it wrong.